Add a note about Wii guitars and whammy bars
After experimentation with my Wiimote and GH5 Genericaster setup on Linux, I've determined that it works fine if you bind the whammy bar to right stick right, specifically. Left goes unused but doesn't seem to affect gameplay.

This particular Wiitar's whammy bar shows up as an axis which only provides positive values under Linux's joystick API, and which has a limited range of mostly-positive values under evdev and SDL2. In both of those cases it isn't possible to bind RPCS3 to the negative values as they never cross the deadzone threshold and would be more or less useless anyway.

Under Windows with WiitarThing this doesn't matter, the default mappings work out of the box, but this note would've saved me hours of fiddling and reading Linux kernel source (lmao yes I went a little too deep), so I aim to pass it along :)