Move AshCentral Room Codes
This part of the guide has done more harm than good. I am tucking it away because there's still use for it. Updated Spanish image with new localization thanks to Lynx.

The images will more than likely be broken so there will likely be a second commit after this one when I notice it's broken.